Add files via upload
Types of packages in Java
As mentioned in the beginning of this guide that we have two types of packages in java.
1) User defined package: The package we create is called user-defined package.
2) Built-in package: The already defined package like java.io.*, java.lang.* etc are known as built-in packages.
Advantages of using a package in Java
These are the reasons why you should use packages in Java:
•	Reusability: While developing a project in java, we often feel that there are few things that we are writing again and again in our code. Using packages, you can create such things in form of classes inside a package and whenever you need to perform that same task, just import that package and use the class.
•	Better Organization: Again, in large java projects where we have several hundreds of classes, it is always required to group the similar types of classes in a meaningful package name so that you can organize your project better and when you need something you can quickly locate it and use it, which improves the efficiency.
•	Name Conflicts: We can define two classes with the same name in different packages so to avoid name collision, we can use packages
Types of packages in Java
we have two types of packages in java.
1) User defined package: The package we create is called user-defined package.
2) Built-in package: The already defined package like java.io.*, java.lang.* etc are known as built-in packages.
What is Class and Object in Java OOPS
What is Class in Java? 
Class are a blueprint or a set of instructions to build a specific type of object
What is Object in Java?
Object is an instance of a class. An object in OOPS is nothing but a self-contained component which consists of methods and properties to make a particular type of data useful.
What is the Difference Between Object and Class in Java?
A Class in object oriented programming is a blueprint or prototype that defines the variables and the methods (functions) common to all Java Objects of a certain kind.
An object in OOPS is a specimen of a class. Software objects are often used to model real-world objects you find in everyday life.

Constructors in Java 
Constructor is a block of code that initializes the newly created object. A constructor resembles an instance method in java but it’s not a method as it doesn’t have a return type
Difference between Constructor and Method
I know I should have mentioned it at the beginning of this guide but I wanted to cover everything in a flow. 
1.	The purpose of constructor is to initialize the object of a class while the purpose of a method is to perform a task by executing java code.
2.	Constructors cannot be abstract, final, static and synchronised while methods can be.
3.	Constructors do not have return types while methods do.